Add option to manually activate a build action
Currently only after a push event the action to build and push
docker image is triggered. For this reason, there are many commit
that just edit dummy lines on a file to force the build process.

This commit add the option to execute the build process manually from
the action tab.